refactor: move backend files into a module so we can use it in notebooks too
