+++ app/home.py
@@ -1,10 +1,9 @@
-from PIL import Image
 import numpy as np
 import streamlit as st
 
-from display import get_image_plot
-import filters
-import images
+from bin.display import get_image_plot
+import bin.filters as filters
+import bin.images as images
 
 st.title("Détection de contours d'images")
 st.write(
@@ -23,9 +22,7 @@ else:
     original_image = images.get_image_by_name(image_name)  # type: ignore
 
 filter_name = st.selectbox("Choisissez votre filtre de contours :", filters.CONTOURS_FILTERS_LIST)
-threshold_type = st.selectbox(
-    "Choisissez votre méthode de seuillage :", ["Arbitraire", "Quantile"]
-)
+threshold_type = st.selectbox("Choisissez votre méthode de seuillage :", ["Arbitraire", "Quantile"])
 
 match threshold_type:
     case "Arbitraire":
