+++ bin/display.py
